typetools/checker-framework 3.26.0 release (#342)

+++ checker/bin/wpi-many.sh
@@ -73,6 +73,13 @@ else
   has_java17="yes"
 fi
 
+# shellcheck disable=SC2153 # testing for JAVA18_HOME, not a typo of JAVA_HOME
+if [ "${JAVA18_HOME}" = "" ]; then
+  has_java18="no"
+else
+  has_java18="yes"
+fi
+
 if [ "${has_java_home}" = "yes" ] && [ ! -d "${JAVA_HOME}" ]; then
     echo "JAVA_HOME is set to a non-existent directory ${JAVA_HOME}"
     exit 1
@@ -92,6 +99,10 @@ if [ "${has_java_home}" = "yes" ]; then
       export JAVA17_HOME="${JAVA_HOME}"
       has_java17="yes"
     fi
+    if [ "${has_java18}" = "no" ] && [ "${java_version}" = 18 ]; then
+      export JAVA18_HOME="${JAVA_HOME}"
+      has_java18="yes"
+    fi
 fi
 
 if [ "${has_java8}" = "yes" ] && [ ! -d "${JAVA8_HOME}" ]; then
@@ -109,8 +120,13 @@ if [ "${has_java17}" = "yes" ] && [ ! -d "${JAVA17_HOME}" ]; then
     exit 1
 fi
 
-if [ "${has_java8}" = "no" ] && [ "${has_java11}" = "no" ] && [ "${has_java17}" = "no" ]; then
-    echo "No Java 8, 11, or 17 JDKs found. At least one of JAVA_HOME, JAVA8_HOME, JAVA11_HOME, or JAVA17_HOME must be set."
+if [ "${has_java18}" = "yes" ] && [ ! -d "${JAVA18_HOME}" ]; then
+    echo "JAVA18_HOME is set to a non-existent directory ${JAVA18_HOME}"
+    exit 1
+fi
+
+if [ "${has_java8}" = "no" ] && [ "${has_java11}" = "no" ] && [ "${has_java17}" = "no" ] && [ "${has_java18}" = "no" ]; then
+    echo "No Java 8, 11, 17, or 18 JDKs found. At least one of JAVA_HOME, JAVA8_HOME, JAVA11_HOME, JAVA17_HOME, or JAVA18_HOME must be set."
     exit 1
 fi
 
@@ -287,7 +303,7 @@ else
     # Don't match arguments like "-J--add-opens=jdk.compiler/com.sun.tools.java"
     # or "--add-opens=jdk.compiler/com.sun.tools.java".
     # shellcheck disable=SC2046
-    grep -oh "\S*\.java" $(cat "${OUTDIR}-results/results_available.txt") | sed "s/'//g" | grep -v '^\-J' | grep -v '^\-\-add\-opens' | sort | uniq > "${listpath}"
+    grep -oh "^\S*\.java" $(cat "${OUTDIR}-results/results_available.txt") | sed "s/'//g" | grep -v '^\-J' | grep -v '^\-\-add\-opens' | sort | uniq > "${listpath}"
 
     if [ ! -s "${listpath}" ] ; then
         echo "${listpath} has size zero"
@@ -301,13 +317,17 @@ else
 
     mkdir -p "${SCRIPTDIR}/.scc"
     cd "${SCRIPTDIR}/.scc" || exit 5
-    wget -nc "https://github.com/boyter/scc/releases/download/v2.13.0/scc-2.13.0-i386-unknown-linux.zip"
+    wget -nc "https://github.com/boyter/scc/releases/download/v2.13.0/scc-2.13.0-i386-unknown-linux.zip" \
+      || (sleep 60 && wget -nc "https://github.com/boyter/scc/releases/download/v2.13.0/scc-2.13.0-i386-unknown-linux.zip")
     unzip -o "scc-2.13.0-i386-unknown-linux.zip"
 
     # shellcheck disable=SC2046
-    "${SCRIPTDIR}/.scc/scc" --output "${OUTDIR}-results/loc.txt" \
-        $(< "${listpath}")
-
+    if ! "${SCRIPTDIR}/.scc/scc" --output "${OUTDIR}-results/loc.txt" $(< "${listpath}") ; then
+      echo "Problem in wpi-many.sh while running scc."
+      echo "  listpath = ${listpath}"
+      echo "  generated from ${OUTDIR}-results/results_available.txt"
+      exit 1
+    fi
     rm -f "${listpath}"
   else
     echo "skipping computation of lines of code because the operating system is not linux: ${OSTYPE}}"
